backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
pandas
openpyxl
pydantic
openai
python-dotenv
python-magic; platform_system != 'Windows'
python-magic-bin; platform_system == 'Windows'
sqlmodel
aiosqlite
